requirements.txt
django==5.2.1
djangorestframework==3.16.0
gunicorn==21.2.0
requests==2.32.3
django-cors-headers==4.3.1 